requirements.txt
fastapi
uvicorn
rio-tiler
requests
numpy
pillow
paramiko
psutil
sse-starlette
httpx